+++ app/src/main/AndroidManifest.xml
@@ -64,32 +64,32 @@
 
 
         <activity
-            android:name="de.tum.in.tca.ticketcheck.component.ticket.EventsActivity"
+            android:name="de.tum.in.tca.ticketcheck.component.ticket.activity.EventsActivity"
             android:configChanges="orientation|screenSize"
             android:label="@string/events_tickets"
             android:launchMode="singleTop">
         </activity>
         <activity
-            android:name="de.tum.in.tca.ticketcheck.component.ticket.TicketScanActivity"
+            android:name="de.tum.in.tca.ticketcheck.component.ticket.activity.TicketScanActivity"
             android:configChanges="orientation|screenSize"
             android:label="@string/app_name"
             android:launchMode="singleTop"
-            android:parentActivityName="de.tum.in.tca.ticketcheck.component.ticket.AdminDetailsActivity"
+            android:parentActivityName="de.tum.in.tca.ticketcheck.component.ticket.activity.AdminDetailsActivity"
             android:screenOrientation="portrait"
             android:theme="@style/AppThemeFullscreen">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="de.tum.in.tca.ticketcheck.component.ticket.AdminDetailsActivity" />
+                android:value="de.tum.in.tca.ticketcheck.component.ticket.activity.AdminDetailsActivity" />
         </activity>
         <activity
-            android:name="de.tum.in.tca.ticketcheck.component.ticket.AdminDetailsActivity"
+            android:name="de.tum.in.tca.ticketcheck.component.ticket.activity.AdminDetailsActivity"
             android:configChanges="orientation|screenSize"
             android:label="@string/ticket_detail"
             android:launchMode="singleTop"
-            android:parentActivityName="de.tum.in.tca.ticketcheck.component.ticket.EventsActivity">
+            android:parentActivityName="de.tum.in.tca.ticketcheck.component.ticket.activity.EventsActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="de.tum.in.tca.ticketcheck.component.ticket.EventsActivity" />
+                android:value="de.tum.in.tca.ticketcheck.component.ticket.activity.EventsActivity" />
         </activity>
     </application>
