+++ app/src/main/java/de/tum/in/tca/ticketcheck/component/generic/activity/BaseActivity.java
@@ -14,7 +14,7 @@ import android.widget.TextView;
 
 import de.tum.in.tca.ticketcheck.R;
 import de.tum.in.tca.ticketcheck.component.generic.drawer.DrawerMenuHelper;
-import de.tum.in.tca.ticketcheck.component.ticket.EventsActivity;
+import de.tum.in.tca.ticketcheck.component.ticket.activity.EventsActivity;
 import de.tum.in.tca.ticketcheck.utils.Const;
 import de.tum.in.tca.ticketcheck.utils.Utils;
