+++ app/src/main/java/de/tum/in/tca/ticketcheck/component/ticket/activity/ConfirmCheckInFragment.java
@@ -1,4 +1,4 @@
-package de.tum.in.tca.ticketcheck.component.ticket;
+package de.tum.in.tca.ticketcheck.component.ticket.activity;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
