+++ app/src/main/java/de/tum/in/tca/ticketcheck/component/ticket/activity/TicketListAdapter.java
@@ -1,4 +1,4 @@
-package de.tum.in.tca.ticketcheck.component.ticket;
+package de.tum.in.tca.ticketcheck.component.ticket.activity;
 
 import android.content.Context;
 import android.view.LayoutInflater;
