+++ app/src/main/res/layout/fragment_confirm_check_in.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".component.ticket.ConfirmCheckInFragment">
+    tools:context=".component.ticket.activity.ConfirmCheckInFragment">
 
     <TextView
         android:id="@+id/name_text_view"
